reconfigure fakes
