--- spec/fake_standard_out.rb
@@ -1,12 +1,0 @@
-class FakeStandardOut
-  attr_reader :log
-
-  def initialize
-    @log = ""
-  end
-
-  def puts(message)
-    @log += message
-    nil
-  end
-end
